Rename 'Test.DPOR.Schedule' to 'Test.DejaFu.Schedule'
